+++ NEWS.md
@@ -1,5 +1,12 @@
 Phan NEWS
 
+?? ??? 2019, Phan 1.2.2 (dev)
+-----------------------
+
+Language Server/Daemon mode:
++ Make code completion immediately after typing `->` and `::` behave more consistently (#2343)
+  Note: this fix only applies at the very last character of a line
+
 18 Jan 2019, Phan 1.2.1
 -----------------------